Feature/138 nachkommastelle keyresult

+++ backend/src/main/java/ch/puzzle/okr/dto/KeyResultDto.java
@@ -13,13 +13,13 @@ public class KeyResultDto {
     private String ownerLastname;
     private ExpectedEvolution expectedEvolution;
     private Unit unit;
-    private Long basicValue;
-    private Long targetValue;
+    private Double basicValue;
+    private Double targetValue;
     private Long progress;
 
     public KeyResultDto(Long id, Long objectiveId, String title, String description, Long ownerId,
             String ownerFirstname, String ownerLastname, ExpectedEvolution expectedEvolution, Unit unit,
-            Long basicValue, Long targetValue, Long progress) {
+            Double basicValue, Double targetValue, Long progress) {
         this.id = id;
         this.objectiveId = objectiveId;
         this.title = title;
@@ -106,19 +106,19 @@ public class KeyResultDto {
         this.unit = unit;
     }
 
-    public Long getBasicValue() {
+    public Double getBasicValue() {
         return basicValue;
     }
 
-    public void setBasicValue(Long basicValue) {
+    public void setBasicValue(Double basicValue) {
         this.basicValue = basicValue;
     }
 
-    public Long getTargetValue() {
+    public Double getTargetValue() {
         return targetValue;
     }
 
-    public void setTargetValue(Long targetValue) {
+    public void setTargetValue(Double targetValue) {
         this.targetValue = targetValue;
     }
