+++ backend/src/main/java/ch/puzzle/okr/models/KeyResult.java
@@ -33,10 +33,10 @@ public class KeyResult {
     private Unit unit;
 
     @NotNull
-    private Long basisValue;
+    private Double basisValue;
 
     @NotNull
-    private Long targetValue;
+    private Double targetValue;
 
     @NotNull
     @ManyToOne
@@ -114,19 +114,19 @@ public class KeyResult {
         this.unit = unit;
     }
 
-    public Long getBasisValue() {
+    public Double getBasisValue() {
         return basisValue;
     }
 
-    public void setBasisValue(Long basisValue) {
+    public void setBasisValue(Double basisValue) {
         this.basisValue = basisValue;
     }
 
-    public Long getTargetValue() {
+    public Double getTargetValue() {
         return targetValue;
     }
 
-    public void setTargetValue(Long targetValue) {
+    public void setTargetValue(Double targetValue) {
         this.targetValue = targetValue;
     }
 
@@ -183,8 +183,8 @@ public class KeyResult {
         private @NotNull User owner;
         private @Size(min = 2, max = 250) ExpectedEvolution expectedEvolution;
         private @NotNull @NotBlank Unit unit;
-        private @NotNull Long basisValue;
-        private @NotNull Long targetValue;
+        private @NotNull Double basisValue;
+        private @NotNull Double targetValue;
         private @NotNull User createdBy;
         private @NotNull LocalDateTime createdOn;
 
@@ -230,12 +230,12 @@ public class KeyResult {
             return this;
         }
 
-        public Builder withBasisValue(@NotNull Long basisValue) {
+        public Builder withBasisValue(@NotNull Double basisValue) {
             this.basisValue = basisValue;
             return this;
         }
 
-        public Builder withTargetValue(@NotNull Long targetValue) {
+        public Builder withTargetValue(@NotNull Double targetValue) {
             this.targetValue = targetValue;
             return this;
         }
